Add files via upload
This analysis is a project I  did under forage Data Analytics and visualization job simulation. The project focused on a social business who majorly involve in content creation.
The problem of the company was their inability to meet up with the high scalability of the large amount of data received which majorly was unstructured. Secondly, the company wanted to go for IPO and needed assistance on how to initiate it successfully.
Finally, to know the top 5 content categories  based on popularity.